pyproject.toml
[project]
name = "petscagent-bench"
version = "0.1.0"
description = "Agentified benchmark for PETSc coding agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.8",
    "dotenv>=0.9.9",
    "litellm>=1.80.5",
    "typer>=0.19.2",
    "uvicorn>=0.37.0",
]
